// frontend/src/services/api/models/EmployeeInDB.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EmployeeInDB = {
    emp_pin?: (string | null);
    emp_ssn?: (string | null);
    emp_role?: (string | null);
    emp_firstname: string;
    emp_lastname?: (string | null);
    emp_username?: (string | null);
    emp_pwd?: (string | null);
    emp_email?: (string | null);
    emp_timezone?: (string | null);
    emp_phone?: (string | null);
    emp_payroll_id?: (string | null);
    emp_payroll_type?: (string | null);
    emp_pin2?: (string | null);
    emp_photo?: (string | null);
    emp_privilege?: (string | null);
    emp_group?: (string | null);
    emp_hiredate?: (string | null);
    emp_address?: (string | null);
    emp_active?: boolean;
    emp_firedate?: (string | null);
    emp_firereason?: (string | null);
    department_id?: (number | null);
    position_id?: (number | null);
    manager_id?: (number | null);
    status?: (string | null);
    id: number;
};